refactor: Break down update.rs into modular components (546 → 188 lines)
Refactored the monolithic update.rs (546 lines, 6 functions, 52+ complexity
indicators) into a well-organized module structure with clear separation of concerns.

**Module Structure:**
```
crates/executor/src/update/
├── mod.rs               # Main UpdateExecutor (188 lines)
├── constraints.rs       # Constraint validation (213 lines)
├── foreign_keys.rs      # FK validation (110 lines)
├── row_selector.rs      # WHERE clause & PK optimization (130 lines)
└── value_updater.rs     # Assignment evaluation (87 lines)
```

**Key Improvements:**

1. **Foreign Key Validation** (`foreign_keys.rs`)
   - Extracted `ForeignKeyValidator` with:
     - `validate_constraints()`: Check parent table references
     - `check_no_child_references()`: Prevent orphaned child rows
   - Preserves dual validation (before PK update, after value update)

2. **Row Selection** (`row_selector.rs`)
   - Extracted `RowSelector` with primary key index optimization
   - Moved `extract_primary_key_lookup()` for O(1) PK lookups
   - Moved `collect_candidate_rows()` for table scan fallback
   - Maintains WHERE CURRENT OF cursor TODO (line 433)

3. **Constraint Validation** (`constraints.rs`)
   - Extracted `ConstraintValidator` with modular validation:
     - `validate_not_null()`: NOT NULL constraints
     - `validate_primary_key()`: PK uniqueness (with index optimization)
     - `validate_unique_constraints()`: UNIQUE constraints (with index optimization)
     - `validate_check_constraints()`: CHECK constraints
   - Each validator independently testable

4. **Value Updates** (`value_updater.rs`)
   - Extracted `ValueUpdater` for assignment evaluation
   - Handles DEFAULT expression evaluation (lines 219-235 preserved)
   - Maintains two-phase semantics (evaluate against original row)
   - Returns (updated_row, changed_columns) for selective index maintenance

5. **Main Executor** (`mod.rs`)
   - Reduced from 546 to 188 lines (66% reduction)
   - Clear orchestration of components
   - Preserved #839 optimization (selective index update via `update_row_selective`)
   - Preserved #841 optimization (cursor-level schema caching via `execute_with_schema`)

**Preserved Optimizations:**
- ✅ #839: Selective index update (only rebuild affected indexes)
- ✅ #841: Cursor-level schema caching
- ✅ Primary key index O(1) lookups
- ✅ Two-phase update semantics

**Testing:**
- ✅ All 42 UPDATE tests pass
- ✅ All executor tests pass
- ✅ No performance regressions
- ✅ Public API unchanged (backward compatible)

**File Metrics:**
- Original: 546 lines, 1 file, 6 functions, 52+ complexity indicators
- Refactored: 728 total lines, 5 files, ~20 functions, <10 complexity per function
- Main orchestration: 188 lines (66% reduction)
- Largest module: 213 lines (constraints.rs, well below 300 target)

Closes #843
